fix(sensor): sync LSM6DSR gyro ODR caches

+++ components/sensor/sensors/sensor_lsm6dsr.h
@@ -80,7 +80,7 @@ typedef struct {
     uint8_t accel_range;
     uint8_t gyro_range;
     uint32_t accel_rate;
-    uint8_t gyro_rate;
+    uint32_t gyro_rate;
     int16_t accel_offset[3];
     int16_t gyro_offset[3];
 } lsm6dsr_priv_t;
